refactoring: Arcade & document
